🛠️ refactor(keymaps): rename terminal keymap module to left-right
