+++ src/keymaps/index.ts
@@ -1,3 +1,3 @@
 export * from './ai/ai-handler';
 export * from './ai/utils';
-export * from './terminal/terminal';
+export * from './terminal/left-right';
