+++ src/keymaps/terminal/left-right.ts
